requirements.txt
cryptography>=42.0.0
Django==5.2.3
djangorestframework==3.16.0
django-allauth>=0.55.0
dj-rest-auth
drf-yasg[swagger-ui]
python-dotenv==1.1.1
psycopg2-binary==2.9.9
PyJWT==2.10.1
requests